refactor(website): split DashboardVisual into a folder structure
DashboardVisual.tsx shell + barrel, with BarChart, DonutChart, and the
extracted KpiCard in components/; the four dashboard types split one-per-
file into types/; and DASHBOARD_VISUAL_DATA in data/. The shared
WidgetCard/WidgetHeader chrome stays in the shell.
