+++ packages/twenty-website/src/sections/product-feature/DashboardVisual/index.ts
@@ -1,0 +1,1 @@
+export { DashboardVisual } from './DashboardVisual';
