+++ packages/twenty-website/src/sections/product-feature/DashboardVisual/types/dashboard-kpi.ts
@@ -1,0 +1,6 @@
+export type DashboardKpi = {
+  label: string;
+  trendDirection: 'down' | 'up';
+  trendPercent: number;
+  value: string;
+};
