+++ packages/twenty-website/src/sections/product-feature/DashboardVisual/types/dashboard-month.ts
@@ -1,0 +1,4 @@
+export type DashboardMonth = {
+  label: string;
+  values: number[];
+};
