+++ packages/twenty-website/src/sections/product-feature/DashboardVisual/types/dashboard-stage-tone.ts
@@ -1,0 +1,1 @@
+export type DashboardStageTone = 'blue' | 'orange' | 'purple' | 'turquoise';
