+++ packages/twenty-website/src/sections/product-feature/DashboardVisual/types/dashboard-stage.ts
@@ -1,0 +1,6 @@
+import { type DashboardStageTone } from './dashboard-stage-tone';
+
+export type DashboardStage = {
+  label: string;
+  tone: DashboardStageTone;
+};
